requirements.txt
pandas
openpyxl
pytesseract
opencv-python
Pillow
numpy